+++ src/index.ts
@@ -1,4 +1,4 @@
-export { default as Button } from "./components/Button";
+export * from "./components/Button";
 export { default as ButtonMenu } from "./components/ButtonMenu";
 export { default as ButtonMenuItem } from "./components/ButtonMenu/ButtonMenuItem";
 export { default as Checkbox } from "./components/Checkbox";
