+++ src/widgets/Nav/Panel.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 import { CloseIcon, ChevronDownIcon } from "../../components/Svg";
-import Button from "../../components/Button";
+import Button from "../../components/Button/Button";
 import Dropdown from "../../components/Dropdown";
 import MenuDropdwn from "./MenuDropdwn";
 import Language from "./icons/Language";
