+++ src/widgets/Nav/UserBlock.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import Button from "../../components/Button";
+import Button from "../../components/Button/Button";
 import useWalletModal from "../WalletModal";
 import { Login } from "../WalletModal/types";
