+++ src/widgets/WalletModal/AccountModal.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "../../components/Button";
+import Button from "../../components/Button/Button";
 import Text from "../../components/Text";
 import { Link } from "../../components/Link";
 import Flex from "../../components/Flex";
